// src/types/vehicle.ts
export type Vehicle = {
  id: string;
  vehicleid: string;
  brand: string;
  model: string;
  year: number;
  vehicletype: string;
  colors: string;
  locations: string[];
  priceperday: number;
  availability: boolean;
  electricvehicle: boolean;
  carimg: string;
  seats: number;
  luggage: number;
  horstpower: string;
  ps: number;
  consumption: string;
  fuel: string;
  geartype: string;
};